+++ src/channel/index.ts
@@ -1,7 +1,7 @@
-export * as RPC08 from './rpc_0_8_1';
 export * as RPC09 from './rpc_0_9_0';
+export * as RPC010 from './rpc_0_10_0';
 // Default channel
-export * from './rpc_0_9_0';
+export * from './rpc_0_10_0';
 export * from './ws/ws_0_9';
 export * from './ws/subscription';
 export { TimeoutError, WebSocketNotConnectedError } from '../utils/errors';